add rspec to gemfile, made test more explicit for create-table test (#12)
* add rspec to gemfile, made test more explicit for create-table test

* moved the line to drop students table into specific test

* added IF EXISTS to table drop query in #create_table test

+++ spec/student_spec.rb
@@ -4,16 +4,19 @@ describe "Student" do
 
   let(:josh) {Student.new("Josh", "9th")}
 
-  before(:each) do
-    DB[:conn].execute("DROP TABLE IF EXISTS students")
-    sql =  <<-SQL
+  before(:each) do |example|
+    unless example.metadata[:skip_before]
+
+      DB[:conn].execute("DROP TABLE IF EXISTS students")
+      sql =  <<-SQL
       CREATE TABLE IF NOT EXISTS students (
         id INTEGER PRIMARY KEY,
         name TEXT,
         grade TEXT
-        )
-    SQL
-    DB[:conn].execute(sql)
+      )
+      SQL
+      DB[:conn].execute(sql)
+    end
   end
 
   describe "attributes" do
@@ -29,7 +32,8 @@ describe "Student" do
   end
 
   describe "#create_table" do
-    it 'creates the students table in the database' do
+    it 'creates the students table in the database', :skip_before do
+      DB[:conn].execute("DROP TABLE IF EXISTS students")
       Student.create_table
       table_check_sql = "SELECT tbl_name FROM sqlite_master WHERE type='table' AND tbl_name='students';"
       expect(DB[:conn].execute(table_check_sql)[0]).to eq(['students'])
